scheduledBatchCreateArticlesで記事の削除も行う #82
